chore: only use system settings for darkMode
this simplify the logic quite a bit
