+++ app/graphql/local-schema.gql
@@ -15,13 +15,6 @@ extend type Query {
   hiddenBalanceToolTip: Boolean!
   price: String # FIXME test only?
   beta: Boolean!
-  darkMode: AppearanceMode!
-}
-
-enum AppearanceMode {
-  light
-  dark
-  system
 }
 
 extend type Account {
