+++ app/hooks/use-darkmode.tsx
@@ -1,0 +1,5 @@
+import { Appearance } from "react-native"
+
+export const useDarkMode = () => {
+  return Appearance.getColorScheme() !== "light"
+}
